[adb-mdns] Unblock main thread during auto-connect attempt.
When a mdns instance is discovered and we try to auto-connect to it,
we block the fdevent thread from processing anything else. This CL
queues all auto-connect attempts on a separate thread to unblock the
main thread.

Bug: 175394720

Test: With paired devices available on the network (wifi debugging
turned on), start adb server and run 'adb mdns services'. Adb client
should not be blocked on the auto-connection attempts that are happening.

Change-Id: I9392ce00ead5c4f25b2bdcb84bbc8491c200773a