Force-clear bonjour cache on tcp disconnects.
When the device does a hard disconnect (e.g. turning Wi-Fi off),
the bonjour daemon won't see the new SRV record until the ttl of
the record expires. We can speed this up by telling the daemon to
recheck the record if we detect a tcp disconnect.

Bug: 200221368
Test: Auto-connect a device, toggle Wi-Fi off/on for tcp disconnect,
then enable wireless debugging again. Adb server should remove the stale
ip/port and retrieve the new one, which will allow the reconnect to succeed.

Change-Id: Ie61ced2705a9eaadfb23da4927fd1562474e06e9