adb: add `shell --quote`.
`adb shell foo` goes through two layers of shell argument parsing, once
on the host, and once on the device. This is very surprising to users
(despite it being identical to ssh's behavior). Add a flag to do this
automatically by wrapping arguments in single quotes, and then escaping
single quotes inside the argument.

Bug: https://issuetracker.google.com/issues/199919702
Test: python3 -m unittest test_device.ShellTest.test_shell_quote
Change-Id: I13723a28754a1ed283d373922d934c04ebeeea24