ADB_LIBUSB Handle device reboot properly.
If the device is rebooted when ADB_LIBUSB=1, the server tries to cancel
a pending write transfer. libusb_cancel_transfer fails, but the result
is ignored. adb server proceeds to wait forever for write_cb to be
called, and it never is.

This change checks the libusb_cancel_transfer result code if the cancel
fails for any reason, it assumes the transfer is invalid for some reason
and frees it.

Bug: 204146629
Test: Steps described in bug.
Change-Id: Id4ceb69994f02c7fcc1dde3c986624265a028697