adb --one-device feature.
Added support for a new server option, --one-device.

When a server is started with that option, it will only communicate with
the USB device serial number or device address specified.

This allows multiple adb servers to run in parallel on a single host,
each communicating only with a single USB device.

example usage:
$ adb --one-device usb:2-4 -P 1138 start-server
$ adb --one-device usb:2-6 -P 1139 start-server
$ adb -P 1138 devices -l
List of devices attached
97221FFBA00016         device usb:2-4 transport_id:1
$ adb -P 1139 devices -l
List of devices attached
13031FDEE0000C         device usb:2-6 transport_id:1

Test: Manually tested as shown above.
Change-Id: Ibc4e3df7ecdefce737291185fa6a84071bff5340