adbd: do not allow `adb root` if ro.adb.allow_root is false
To automate USER build testing, some partners are overriding the
following properties to disable ADB authentication:

  - ro.adb.secure=0
  - ro.debuggable=1

However, when running `adb root`, the adb connection gets
stuck because it lacks the userdebug sepolicy. Adding a new
property "ro.adb.allow_root" and set it to 0 to disallow
`adb root` to avoid this issue for this scenario.

Bug: 208163696
Test: Set above properties and "ro.adb.allow_root=0" then run `adb root`,
      => see "adbd cannot run as root when ro.adb.allow_root is false".
Change-Id: I2723d2dc9cd7e385769325ab9d44e80e1639a120