adb: add receive windowing.
The adb protocol struggles when used over long fat pipes, because the
a writer must wait for the receiver to acknowledge a write before it can
continue writing data. For example, a connection with infinite
bandwidth, but 50ms of latency is limited to 20MBps of throughput.

We still need some sort of acknowledgement, to be able to supply
backpressure. Take an initial stab at addressing this issue by
implementing receive windowing, which lets the writer know how many
more bytes it can write. Repurpose the unused field in A_OPEN to set the
initial window size, and add a payload to A_OKAY packet that
acknowledges a specified number of bytes.

Test: with adbd running on a separate machine connected with 10GBe:
  [jmgao@atlas:~]# sudo tc qdisc replace dev lan root netem delay 50ms
  [jmgao@atlas:~]$ ./adb kill-server && ADB_DELAYED_ACK=0 ./adb connect 192.168.0.152:5555 && ./adb push 1g.bin 1g.bin
  * daemon not running; starting now at tcp:5037
  * daemon started successfully
  connected to 192.168.0.152:5555
  1g.bin: 1 file pushed, 0 skipped. 19.1 MB/s (1073741824 bytes in 53.685s)

  [jmgao@atlas:~]$ ./adb kill-server && ADB_DELAYED_ACK=1 ./adb connect 192.168.0.152:5555 && ./adb push 1g.bin 1g.bin
  * daemon not running; starting now at tcp:5037
  * daemon started successfully
  connected to 192.168.0.152:5555
  1g.bin: 1 file pushed, 0 skipped. 475.4 MB/s (1073741824 bytes in 2.154s)
Change-Id: I6498d18e5ab4d866a81bf4d5c3d65cebb579c059