Report device connection errors when installing
When installing an APEX and there are device connection issues, it's
expected to see an error related to device connection and not a spurious
message about install mode compatibilities.

Before installing, there is a process to determine the mode that will be
used. Part of this queries the device capabilities but ignores any
connection issues in the process, falling back to some common
denomincator defaults. This change makes those connection issues fail
the command and print the associated error.

The error was previously ignores as pre-N devices did not support
features. But seeing how the features are default empty and added based
on the banner parsing, I don't see how that could be a problem.

Bug: 191621554
Test: atest adb_test
Test: try installing APEX without connected device and see error
Change-Id: I25a182c3bc85b18607cfe1ced1b92069f20a05ef