Fix O_NONBLOCK setting.
You can't set O_NONBLOCK with F_SETFD. You need F_SETFL for that.
Unfortunately, the kernel will silently no-op invalid flags in F_SETFD,
so you'll never know.

It looks like this code is only ever used in cases where we can just
create the socket with the right flags anyway, so let's just do that
instead of try to fix this by doing a read-modify-write on the flags.

Test: treehugger
Change-Id: I77e3a181ee0966d85d8d885054fc11891278ab33