Make adb stream-aware
Bug: 225217720

Make adb exit if it can no longer write to the output stream for
logcat.  This allows you to put adb in a pipe and have it exit
properly if the downstream process exits.

The change is implmented with a '--exit-on-write-error' global switch.
Without the switch, adb behaves in legacy fashion.  With the switch,
several commands (specifically, "logcat") will exit as soon as the
downstream pipe closes.

Manually tested the feature with a build that generated the string
"READY" periodically.

This command never exits:
  adb logcat | grep -q READY

This command exits
  adb --exit-on-write-error logcat | grep -q READY

Test: * presubmit
 * adb logcat (with and without the new switch)
 * adb bugreport
Change-Id: Ifdba393f9f78ed05fd789101fc983e2bb9cca774