remount: Check external command accessible before executing
adb remount, disable-verity, enable-verity are delegated to external
commands, however these executables can be inaccessible if adbd is not
running as root or device is user build. When this happens a cryptic
error message would be printed:

> adb remount
/system/bin/sh: remount: inaccessible or not found

Improve this error message by checking the command accessibility
before executing. If command is inaccessible, hint the user to retry
with adb root on debuggable build.

Test: adb remount / disable-verity / enable-verity
Change-Id: Ib66a5bc9462e327c8a502ba527d55a0a018d6e74