Revert "ADB: Document ConnectionState"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/3cbd6603-80d7-4194-9f3e-beba11995cdc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id3ca2dad7f870de38b6f29c0586210576e1e5f37