Revert "ADB: Document ConnectionState"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/3cbd6603-80d7-4194-9f3e-beba11995cdc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia721c694c3772cfc1191d1222ac4b01a8640c893