ADB: Delete test framework interrupt trick
The test framework features a fd to interrupt the (e)poll(3) call in
the fdevent system. This was created back in aosp/202177 at a time
where the fdevent did not have an interrupt mecanism.

Since then, fdevent has an Interrupt() method which is called automagi-
cally when lambda are added to the runloop via
fdevent_run_on_main_thread which made the test fd interrupt redundant.

Test: Unit tests and integration tests.
Bug: NA
Change-Id: I2500209beddcb98ea2e9aabc124e97aaf54de24a