Revert "Switch to dynamic linking to improve test coverage metri..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/01539951-27a2-4a82-b46e-1721c06b9813). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iec311a696c58107f0e2284067da58ac364d0bd83