Revert "Coverage enhanement for IOVector."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/414a0b60-b814-44d0-a0a6-e5524abda171). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibb6fa7f0355bc0e84184fa1a42d1933c950a3fe1