Revert "Coverage enhanement for IOVector."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/eb4f5bf8-0e58-4f27-b1d0-3bbf0f35031b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I286997b13073fbf6adad0b2ee887f4147b0e11ae