Revert "Fix test for clang update."
Revert "Update clang version to clang-r475365"

This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/14b7a26b-8e68-463d-8625-12a6435491d7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibfb3d59e3803254f5f527f6717a0403cd928b6da