Revert "Include module dependencies under `data`"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bf74eb17-b133-4219-b7f7-9bbe084395ac). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8bacb9209c1e8ad2134c43a34ea661f102dee11b