Revert "Test out fastdeploy test in test mapping"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4bd10b3d-74a5-48d8-af39-21ffe630ebdc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I95b21dd3ec62eb355dd719ac5e49f204449c1bc9