Revert "Test out fastdeploy test in test mapping"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e32c68f6-296c-4652-a883-c6b120f4e1ad). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I181dde8de2ee579e1d6073c9c1c56aa4ccf6f015