Revert "Test out fastdeploy test in test mapping"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5fd3af1d-51bb-4830-a937-1a34f741c67d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ica502623b6ab1c094d927fa8925e5139db0ea5c3