Revert "Test out fastdeploy test in test mapping"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7b7a18f8-ef22-481d-a97f-3f36f8f9be3e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I276e743187d612cd1020ccd65087109a83b258ab