Revert "Test out fastdeploy test in test mapping"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3a332e80-2310-42f3-bfa0-0d9de067b400). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0816a6e5ea2f059ae271a56f1f94a628f9fec153