Re-launch adb server with one-device option.
Problem:
If an adb server crashes on a `one_device_required` system, it won't be
restarted automatically by the next command as the invocation will miss
the `--one-device` flag.

Solution:
If, on a `one_device_required` system, `-s` serial is provided in the
command, use the serial and attempt to re-start the adb server with the
`--one-device` flag when necessary.

Test:
Steps to manually test the adb binary, with example commands.
On a device with "/etc/adb/one_device_required" file present, start an
adb server with `--one-device` option:
$ adb --one-device 0ab70e96d293 -P 27761 start-server

Kill the server:
$ adb -s 0ab70e96d293 -P 27761 kill-server

Run another adb command with the serial (-s) provided:
$ adb -s 0ab70e96d293 -P 27761 devices
The above command should successfully start an adb server.

Bug: b/283038263

Change-Id: Ie5dd49c95e5db59203a2faf78b438accfef88d91