Make default mDNS backend platform dependant
Problem: Client must start adb with special env variable to make it work
on Linux (there is usually no Bonjour service there). If adb server is
retarted by the user, ADB_MDNS_OPENSCREEN is not set and ADB Wifi cease
to function.

Solution: Make openscreen the default backend on Linux. Use Bonjour for
Windows and Darwin (this is the previous behavior).

Test: NA
Bug: 292850997
Change-Id: I87348b3206d5d1a1695e2f93a4162c02e2f989db