Fix crash (Abort trap: 6 when unplugging phone)
Problem: adb server with libusb backend crashes when unplugging a device using an USB3 cable (macosx only). It seems we are too aggressive in checking for terminated_ since the case is handled anyway below in MaybeCleanup.

Solution: Remove agressive trap. adb server doesn't crash if the trap
is removed and looks like it is working normally.

Fixes: 308190337
Test: NA
Change-Id: I8ba10f910b6423e467a1b442aeba1e31c7a6610a