Gracefully close smartsocket upon exit
The function in charge of closing the smartsockets, close_smartsockets,
is not looking for the right field. Upon startup, install_listener
is called and result in a listener create with:

{
  local_name = "tcp:5037";
  connect_to = "*smartsocket*";
}

The field to lookup should be `connect_to`, not `local_name`. The
current implementation is "leaking" the smartsocket. In practice, this
is not a problem because the OS will close it anyway but it is really
confusing to read as a developer.

Test: NA
Change-Id: I020bf57fc81b01b6751cdc37962681b6c4cefda7