Revert "adb: Do not use fs_config unless we are root."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5ce507b2-ad2a-450e-ba96-374f3a3e4bd5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5f31458b18840998214e482766c9f8bd9f72a7cb