Revert "adb: Do not use fs_config unless we are root."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2a83b18c-5fce-44b7-a359-2d8149f80563). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I87aa20c6ef8506f9c92645273726df8993b54ff0