adb: Create a ProgressCallbacks class.
With "adb tombstone" the output on stdout is intended to be read by
not only humans but also other programs. That means that we don't
want files (i.e. tombstones) pulled from the target causing progress
updates to be written to stdout and confusing the other program.

With this change we allow the user of SyncConnection to customize the
progress output. For existing commands like "adb push", "adb pull",
"adb sync" a LegacyProgressCallbacks class is added that leaves
their behavior unchanged, although they morally ought to be printing
to stderr as well. But the intent is that new commands like "adb
tombstone" that copy files to/from the target as an implementation
detail and need to control what is written to stdout can use the
base class which implements behavior suitable for programmatic usage
of SyncConnection (only prints errors and warnings, and prints them
to stderr), or derive from it and implement the behavior that makes
sense for them.

Bug: 328531087
Change-Id: I80506ce16c2b4cc5955bfc5ab88eb46078ed1446