adb: Add "adb tombstone" subcommand.
Downloading tombstones after a crash and symbolizing them is a common
part of the developer workflow for platform developers. Especially
with MTE and GWP-ASan, there is useful debugging information contained
in the tombstone that may be useful to developers. It is possible to
download the tombstones manually and symbolize them with stack but
this takes some manual effort and for the memory tools the allocation
and deallocation traces will not be displayed in line with the rest
of the information in the tombstone requiring developers to consult
both the tombstones and the stack output. Therefore add a subtool of
adb called "adb tombstone" that automatically downloads and symbolizes
the tombstones and writes them to disk on the host.

From the usage message:

usage: adb tombstone [OPTION]

Downloads tombstones from the device and symbolizes them.
Paths to symbolized tombstones are written to stdout.
Requires Android 15 or later, or a rooted device.

Arguments:
  -h, --help                   print this message
  --debug-file-directory PATH  specify the path to a symbols directory;
                               may be specified multiple times;
                               must contain a .build-id subdirectory conforming to
                               https://fedoraproject.org/wiki/RolandMcGrath/BuildID
                               (default: $ANDROID_PRODUCT_OUT/symbols if set)
  -o DIR                       write tombstones to specified directory
                               (default: current directory)
  -f                           monitor for new tombstones after downloading
                               existing tombstones instead of exiting;
                               may not be reliable before Android 15
  -n N                         download N most recent existing tombstones
                               (default: 0 if -f specified, otherwise 3)

Bug: 328531087
Change-Id: Id8c452002ad02f9d3e640afdb22f23a53f725eb3