Fix another adb module backwards compatibility issue.
It turns out that some tests have the following expectation:
- old system
- new modules
- adb push to /data/local/tmp as root
- files are expected to retain host permissions

This breaks because now getuid() == 0 passes but the old version of
adbd_fs_config will override the permission to 644. Fix it by moving
the getuid() check back into adbd_fs_config and adding a version
check in the caller. If the system version is old, we follow the pre-
https://r.android.com/2980341 logic.

Bug: 341171960
Change-Id: Ie92e78be8ac5731b15361ab05418e8950c74ee99