Make sure USB interfaces are released on shutdown
On Windows libusb, not releasing USB interfaces upon shutdown leaves them in unstable state. If the server is restarted within 0-5 seconds,
writing to the endpoint can fail.

Performing a cleanup and destroying all transport/connections upon
shutdown avoid this issue.

Linux libusb does not seem to exhibit this problem but it is not a
bad idea to consistently cleaup upon shutdown on all platforms.

Macos has always be cleaning up (both native and libusb), and native
Windows/Linux backend are left unaffected since they are set to be
deprecated soon (I hope).

Test: NA
Bug: NA
Change-Id: I3324b1dc97392373086ab420b39c9ccf616c99fa