Revert "Fix server-status: Fail if string too short"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5e065b1f-152f-43fd-95cd-6257d7b3eabe). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idddc51b052bb63990b37a82ac182ec0ceba969bb
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com