Revert "Fix server-status: Fail if string too short"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4f5465e3-9850-4aaa-87ad-2d7d9796eb43). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idb0573e3b8af758ce1243ec76e04b62bad7d85f0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com