Revert "Fix server-status: Fail if string too short"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/46e34bc3-d577-49be-9c77-9d989c3addfc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie40d34065c39c9612e86556c37fe19271f0c6d5d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com