Revert "Fix ADB Wifi (broken by aosp/2909778)"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d92cba89-eddd-4ec8-924f-ecef14b287b9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5b45ae3cc8a74801abd4b9c137089e151974a99b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com