adb: add `shell --quote`
Summary: `adb shell foo` goes through two layers of shell argument parsing, once
on the host, and once on the device. This is very surprising to users
(despite it being identical to ssh's behavior). Add a flag to do this
automatically by pre-pending \ to single quotes and \\\ to double quotes.

Bug: https://issuetracker.google.com/issues/199919702

Test: python3 -m unittest test_device.ShellTest.test_shell_quote_single

Change-Id: I92a661b84d7ffd91ed70e80629094af017b146ae
Signed-off-by: Abdelrahman Daim <adaim@meta.com>